backend/user/requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
click==8.2.1
dotenv==0.9.9
fastapi==0.115.12
h11==0.16.0
idna==3.10
packaging==25.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
redis==6.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
alembic==1.14.0
psycopg2-binary==2.9.10
